--- TourMate/TourMate/Frontend/Plan/Views/PlansViews/PlansNewCalendarDayView.swift
@@ -1,199 +1,0 @@
-//
-//  PlansDayView.swift
-//  TourMate
-//
-//  Created by Rayner Lim on 28/3/22.
-//
-
-import SwiftUI
-
-struct PlansNewCalendarDayView: View {
-    @ObservedObject var viewModel: PlansViewModel
-
-    let date: Date
-    let plans: [Plan]
-    let dateFormatter: DateFormatter
-    let hourHeight: Float
-
-    let onSelected: ((Plan) -> Void)?
-
-    @State var planIdToSize: [String: CGSize] = [:]
-    @State var planIdToOffset: [String: CGPoint] = [:]
-    @State var minWidth: CGFloat = 0
-    @GestureState var offset = CGSize.zero
-
-    private let viewModelFactory = ViewModelFactory()
-
-    init(viewModel: PlansViewModel,
-         date: Date,
-         plans: [Plan] = [],
-         onSelected: ((Plan) -> Void)? = nil,
-         hourHeight: Float = 64) {
-
-        self.viewModel = viewModel
-        self.date = date
-        self.plans = plans
-        self.onSelected = onSelected
-        self.dateFormatter = DateFormatter()
-        self.dateFormatter.timeStyle = .short
-        self.hourHeight = hourHeight
-    }
-
-    // TODO: use DateFormatter
-    func getHourString(for hour: Int) -> String {
-        if hour == 0 || hour == 24 {
-            return "12 AM"
-        } else if hour < 12 {
-            return "\(hour) AM"
-        } else if hour == 12 {
-            return "12 PM"
-        } else {
-            return "\(hour - 12) PM"
-        }
-    }
-
-    func getYPosition(for plan: Plan) -> (start: Float, end: Float) {
-        let startDateComponents = Calendar
-            .current
-            .dateComponents(in: plan.startDateTime.timeZone, from: plan.startDateTime.date)
-        let startDate = Calendar.current.date(from: DateComponents(year: startDateComponents.year,
-                                                                   month: startDateComponents.month,
-                                                                   day: startDateComponents.day))!
-        let endDateComponents = Calendar
-            .current
-            .dateComponents(in: plan.endDateTime.timeZone, from: plan.endDateTime.date)
-        let endDate = Calendar.current.date(from: DateComponents(year: endDateComponents.year,
-                                                                 month: endDateComponents.month,
-                                                                 day: endDateComponents.day))!
-        var start: Float = 0
-        var end: Float = 24 * hourHeight
-        if startDate == date {
-            start = (Float(startDateComponents.hour ?? 0) + Float(startDateComponents.minute ?? 0) / 60) * hourHeight
-        }
-        if endDate == date {
-            end = (Float(endDateComponents.hour ?? 0) + Float(endDateComponents.minute ?? 0) / 60) * hourHeight
-        }
-
-        return (start: start, end: end)
-    }
-
-    func getHeight(for plan: Plan) -> Float {
-        let yPosition = getYPosition(for: plan)
-        return yPosition.end - yPosition.start
-    }
-
-    func calculateOffsets() {
-        var planIdRect: [(planId: String, rect: CGRect)] = []
-
-        // Get initial rects
-        for plan in plans {
-            let origin = CGPoint(x: CGFloat(0.0), y: CGFloat(getYPosition(for: plan).start))
-            let size = planIdToSize[plan.id] ?? CGSize(width: 0, height: 0)
-            let rect = CGRect(origin: origin, size: size)
-            planIdRect.append((planId: plan.id, rect: rect))
-        }
-
-        // Set Y offset for first plan
-        if !planIdRect.isEmpty {
-            let firstPlanIdRect = planIdRect[0]
-            planIdToOffset[firstPlanIdRect.planId] = CGPoint(x: 0, y: firstPlanIdRect.rect.origin.y)
-            minWidth = firstPlanIdRect.rect.width
-        }
-
-        // Calculate and set X, Y offsets to not overlap
-        for i in 1 ..< planIdRect.count {
-            let currentPlanIdRect = planIdRect[i]
-            let currentPlanId = currentPlanIdRect.planId
-            let currentRect = currentPlanIdRect.rect
-
-            for j in 0..<i {
-                let prevPlanIdRect = planIdRect[j]
-                let prevPlanId = prevPlanIdRect.planId
-                let prevRect = prevPlanIdRect.rect
-
-                // If Y position overlaps, offset to the right
-                let prevEndY = prevRect.origin.y + prevRect.size.height
-                let currentStartY = currentRect.origin.y
-                if currentStartY < prevEndY {
-                    var prevEndX = prevRect.origin.x + prevRect.size.width
-                    if let offset = planIdToOffset[prevPlanId] {
-                        prevEndX += offset.x
-                    }
-                    planIdToOffset[currentPlanId] = CGPoint(x: prevEndX, y: currentStartY)
-                    minWidth = max(minWidth, prevEndX + currentRect.width)
-                } else if planIdToOffset[currentPlanId] == nil {
-                    planIdToOffset[currentPlanId] = CGPoint(x: 0, y: currentStartY)
-                }
-            }
-        }
-    }
-
-    var body: some View {
-        ScrollView {
-            HStack {
-                VStack(alignment: .leading, spacing: 0) {
-                    ForEach(0...24, id: \.self) { hour in
-                        VStack {
-                            Text(getHourString(for: hour))
-                                .font(.caption)
-                            Spacer()
-                        }
-                        .frame(height: CGFloat(hourHeight))
-                    }
-                }
-                ZStack {
-                    ScrollView(.horizontal) {
-                        VStack(alignment: .leading, spacing: 0) {
-                            ZStack(alignment: .topLeading) {
-                                ForEach(plans, id: \.id) { plan in
-                                    HStack {
-                                        PlanBoxView(plansViewModel: viewModel,
-                                                    plan: plan, date: date)
-                                            .onTapGesture {
-                                                if let onSelected = onSelected {
-                                                    onSelected(plan)
-                                                }
-                                            }
-                                            .gesture(DragGesture(minimumDistance: 0, coordinateSpace: .global)
-                                                            .updating($offset, body: {t, state, _ in
-                                                                // print(t.translation)
-                                                                state = t.translation
-                                                        }))
-                                            .frame(maxWidth: UIScreen.screenWidth / 3,
-                                                   minHeight: CGFloat(getHeight(for: plan)),
-                                                   alignment: .topLeading)
-                                            .readSize { size in
-                                                planIdToSize[plan.id] = size
-                                                calculateOffsets()
-                                            }
-                                            .buttonStyle(PlainButtonStyle())
-                                            .background(
-                                                RoundedRectangle(cornerRadius: 8)
-                                                    .fill(Color.primary.opacity(0.25))
-                                            )
-                                            .offset(x: CGFloat(planIdToOffset[plan.id]?.x ?? 0) + offset.width,
-                                                    y: CGFloat(planIdToOffset[plan.id]?.y ?? 0) + 6 + offset.height)
-                                        Spacer()
-                                    }
-                                }
-                            }
-                            Spacer()
-                        }
-                        .frame(minWidth: minWidth)
-                    }
-                    VStack(alignment: .leading, spacing: 0) {
-                        ForEach(0...24, id: \.self) { _ in
-                            VStack(alignment: .leading, spacing: 0) {
-                                Divider()
-                                Spacer()
-                            }
-                            .frame(height: CGFloat(hourHeight))
-                            .offset(y: 6)
-                        }
-                    }
-                }
-            }
-            .padding()
-        }
-    }
-}
