+++ TourMate/TourMate/Frontend/Plan/Views/PlansViews/PlansView.swift
@@ -33,12 +33,6 @@ struct PlansView: View {
                 .pickerStyle(.segmented)
 
                 NavigationLink {
-                    PlansNewCalendarView(viewModel: plansViewModel, onSelected: onSelected)
-                } label: {
-                    Label("Calendar", systemImage: "calendar.day.timeline.left")
-                }
-
-                NavigationLink {
                     PlansMapView(viewModel: plansViewModel, onSelected: onSelected)
                 } label: {
                     Label("Map", systemImage: "map.fill")
